testsuite: Get rid of dupe "auxreg_write_read"
There were 2 "auxreg_write_read" macros lurking in tests codes.
One in "check_mpu.S" and the other in "macros.inc". This was
causing the assembler to choke:

------------8<--------------

.../tests/tcg/arc/check_mpu.S: Assembler messages:
.../tests/tcg/arc/check_mpu.S:38: Error: Macro 'auxreg_write_read'
was already defined make: *** [Makefile:97: check_mpu.o] Error 1

------------>8--------------

Since that macro is not used anywhere but the "check_mpu.S", I
got rid of the one in "macros.inc".