gdb: Add FPU registers
Support for the following registers are added:

f0 ... f31              rw
fp_build                r
fp_ctrl                 rw
fp_status               rw

Although it is possible to write to the registers, it does not
come with the functionalities expected. For instance, writing to
CIV (clear invalid) bit of fp_status will NOT result in clearing
its IV bit. Those are for when the FPU implementation is done.

One can run QEMU with FPU support with:
$ qemu-system-arc64 ... -global cpu.fpu=true ...