helpers: Swapped helper level ld/st with ld_ra/st_ra counterparts
When there is an exception (i.e. SIGSEV) in a load/store inside a
helper, the CPU needs to be given a propper return address.

The currently used functions end up using 0 as a return address,
meaning should the CPU try to return to that address it will fail
again.

This change makes llocks/sconds variants in particular throw
exceptions properly

Signed-off-by: Bruno Mauricio <brunoasmauricio@gmail.com>