target/arc: v3: rewrite translation functions for memory operations
Similarly to the v2 patch, this one replaces the arc_gen_*() functions
by hand-written code. The single and double versions of the instructions
are based on semfunc-v2.c; in addition to those, the v3-specific long
versions are also implemented.

These changes have been validated by running check-tcg and gcc.dg/dg.exp
tests for the arcv3 architecture, as well as booting Linux for the hs6x
CPU.

Signed-off-by: Artemiy Volkov <artemiy@synopsys.com>